pyproject.toml
[tool.poetry]
name = "library-explore"
version = "0.1.0"
description = ""
authors = ["Nikola Nikolovski <nikolovski.nikola42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pymongo = ">=4.9,<4.10"
qdrant-client = "^1.12.1"
requests = "^2.32.3"
huggingface-hub = "^0.26.2"
openai = "^1.55.0"
gitpython = "^3.1.43"
bs4 = "^0.0.2"
markdown = "^3.7"
transformers = "^4.46.3"
langchain-text-splitters = "^0.3.2"
torch = "^2.5.1"
datasets = "^3.1.0"
einops = "^0.8.0"
readability-lxml = "^0.8.1"
lxml = {extras = ["html-clean"], version = "^5.3.0"}
markdownify = "^0.14.1"
websockets = "^14.1"
motor = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
